Geridönüş neticesinde gerekli düzenlemeler yapıldı
